chore: extract load-postgres-changes suite into src/suites/load-postgres-changes.ts
Part of REAL-1037.
